+++ libraries/DS18B20_RT/DS18B20.h
@@ -2,7 +2,7 @@
 //
 //    FILE: DS18B20.h
 //  AUTHOR: Rob.Tillaart@gmail.com
-// VERSION: 0.1.14
+// VERSION: 0.1.15
 //    DATE: 2017-07-25
 // PUPROSE: library for DS18B20 temperature sensor with minimal footprint
 //     URL: https://github.com/RobTillaart/DS18B20_RT
@@ -19,7 +19,7 @@
 //
 
 
-#define DS18B20_LIB_VERSION     (F("0.1.14"))
+#define DS18B20_LIB_VERSION     (F("0.1.15"))
 
 #include <OneWire.h>
 
@@ -27,7 +27,7 @@
 #define DEVICE_DISCONNECTED     -127
 #define DEVICE_CRC_ERROR        -128
 
-//  config codes
+//  configuration codes
 #define DS18B20_CLEAR           0x00
 #define DS18B20_CRC             0x01
 
@@ -41,6 +41,7 @@ class DS18B20
 public:
   explicit  DS18B20(OneWire * ow);
   bool      begin(uint8_t retries = 3);
+  bool      isConnected(uint8_t retries = 3);
 
   void      requestTemperatures(void);
   bool      isConversionComplete(void);